Update variable types and function signatures

+++ data_structure/node.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Node struct {
-	Value    int16   `json:"v"`
+	Value    int32   `json:"v"`
 	Children []*Node `json:"k"`
 }
 
-func NewNode(value int16) *Node {
+func NewNode(value int32) *Node {
 	temp := new(Node)
 	temp.Value = value
 	return temp
@@ -36,7 +36,7 @@ func (this *Node) Convert() interface{} {
 
 }
 
-func (this *Node) AddLang(lang int16) {
+func (this *Node) AddLang(lang int32) {
 	if this.Value&lang == 0 {
 		this.Value += lang
 	}
@@ -55,7 +55,7 @@ func (this *Node) HasChildren() bool {
 	return len(this.Children) != 0
 }
 
-func SameKey(val1, val2 int16) bool {
+func SameKey(val1, val2 int32) bool {
 	//fmt.Printf("%s,%s,%s", string(Decode(val1)), string(Decode(val2)), "\n")
 	return val1&((1<<8)-1) == val2&((1<<8)-1)
 }
@@ -64,7 +64,7 @@ func (this *Node) GetChild(value rune) (*Node, error) {
 
 	for _, node := range this.Children {
 
-		if SameKey(node.Value, int16(value)) {
+		if SameKey(node.Value, int32(value)) {
 			return node, nil
 		}
 	}
@@ -95,7 +95,7 @@ func (this *Node) Add(node *Node) {
 
 // The `CheckWord` function is used to check if a given word exists in the tree and if its associated
 // value matches a given test value.
-func (this *Node) CheckWord(w string, lang int16) bool {
+func (this *Node) CheckWord(w string, lang int32) bool {
 
 	next := this
 	if len(w) < 3 {
